fix files after fetching the last version of branch
