+++ src/domains/reminder/component/reminder-card/reminder-card.js
@@ -1,26 +1,45 @@
 import { Link } from "react-router-dom";
-import useDeleteReminder from "../../services/use-delete-reminder";
+import deleteReminder from "../../services/delete-reminder";
+import { useTranslation } from "react-i18next";
 import "../reminder-card/reminder-card.scss";
 import Icon from "./icon";
 import "./selection.json";
 
 const ReminderCard = (props) => {
-	const deleteReminder = useDeleteReminder;
+	const { t } = useTranslation();
 
 	const handleDelete = (id) => {
 		deleteReminder(id).then(response => console.log(response));
 	};
+
+	const handleFormatDate = (date) => {
+		return (date?.slice(0,10) ?? "");
+	};
+
 	return(
 
 		<div className="reminderCard" key={props.id}>
-			
+
 			<Link className="reminderCard__info" to={`/reminder/${ props.id }`}>
 				<img src="https://via.placeholder.com/100x100?text=image" alt="reminder poster"/>
+				
 				<div className="reminderCard__info--elements">
 					<h2>{props.name}</h2>
-					<div>Type: {props.type}</div>
-					<div>Date of release: {props.date ? props.date.slice(0,10) : ""}</div>
-					<div>Comment: {props.comment}</div>
+					
+					<div>
+						<span>{t("reminder:card.type")} </span>
+						<span>{props.type}</span>
+					</div>
+
+					<div>
+						<span>{t("reminder:card.release-date")} </span>
+						<span>{handleFormatDate(props.date)}</span>
+					</div>
+
+					<div>
+						<span>{t("reminder:card.comment")} </span>
+						<span>{props.comment}</span>
+					</div>
 				</div>
 			</Link>
 			
@@ -36,3 +55,4 @@ const ReminderCard = (props) => {
 };
 
 export default ReminderCard;
+
