+++ src/domains/reminder/component/searchbar/searchBar.js
@@ -3,31 +3,26 @@ import Icon from "./icon";
 import "./selection.json";
 import "./searchbar.scss";
 
-const SearchBar = (props) => {
-	const [search, setSearch] = useState("");
+const SearchBar = ({ setSearch }) => {
+	const [tempSearch, setTempSearch] = useState("");
 	
 	const handleClickSearch = async () => {
-		console.log("search enter");
-		const result = await fetch(`http://localhost:3001/reminder/search/${ search }`);
-		const searchedData = await result.json();
-		return searchedData;
+		setSearch(tempSearch);
 	};
 
-	const handleChangeSearch = (e) => {
-		console.log("value > ", e.target.value);
-		setSearch(e.target.value);
-	};
+	const handleChangeSearch = (e) => setTempSearch(e.target.value);
 	
 	const handleEnter = ({ code }) => {
 		if(code === "Enter") {
-			handleClickSearch();
+			setSearch(tempSearch);
 		};
 	};
 
 	return (
 		<>
 			<div className="search">
-				<input type="text" placeholder="search..." className="search__input" onKeyPress={handleEnter} value={search} onChange={handleChangeSearch}></input>
+				<input type="search" placeholder="search..." className="search__input" onKeyPress={handleEnter} 
+					value={tempSearch} onChange={handleChangeSearch}></input>
 				<button className="search__button" onClick={handleClickSearch}>
 					<Icon icon="magnifying-glass" className="search__icon" />
 				</button>
