+++ src/domains/reminder/pages/reminder-list-page/reminder-list-page.js
@@ -42,7 +42,6 @@ const ReminderListPage = () => {
 		}));
 	};
 
-
 	const typesReminder = ["book", "game", "movie", "all"];
 
 	const handleReminderNotFound = () => {
